[housekeeping] convert gif files to png
My belief is that the pngs are in all cases smaller in filesize than the gifs
were.  That may only be because I've stripped pointless cruft from the png and
if i'd done the same with the gifs they might have been smaller.

But png is a better format, and having done the work, it's a net win both in
file size and format.