[housekeeping] Spring clean book review pages
Convert uris to absolute, but drop the domain

    - <link rel="stylesheet" href="http://edinburgh.pm.org/css/edinburgh-pm-org.css">
    + <link rel="stylesheet" href="/css/edinburgh-pm-org.css">

  and

    - <a href="http://edinburgh.pm.org/">
    + <a href="/">edinburgh.pm</a>

This means that you can checkout the git repo and run it locally on some other
domain and it'll still work right.

---

Remove <base="http://edinburgh.pm.org/">

I can see no benefit to having it, and it scuppers trying to work on a local
checkout.

---

Change "Page maintained by" links from the admin email to the github uri

    -Page maintained by <a href="mailto:owner-edinburgh-pm-list@pm.org">Rory Macdonald</a>
    +Page maintained via <a href="https://github.com/minty/edinburgh.pm">github.com/minty/edinburgh.pm</a>

---

Drop the "Last updated" text.  I wasn't sure if replacing gifs with pngs and
the other minor housekeeping here really warranted an update to the "Last
updated" timestamp.  As Aaron pointed out, because pages are now tracked in
git, those that want to can find that information (and a lot more detail)
there.