[build] convert html files to be template based
pm.org doesn't allow Apache SSIs or any dynamic content.

So the plan is to instead create a template file for each html page we want,
and then have script/build_static search for templates, process them and save
the output as a .html

This then allows us to use concept of a "WRAPPER" page to contain all the
boilerplate.

At the cost of having to run the script/build_static in order to see changes.

I can't see a better way, suggestions welcome.

Then remove the original .html files, and update .gitignore so that it doesn't
bitch about the generated files.