Add links back to e.pm.org homepage
I've killed the old ".navigation" crumbtrail.  Given we only have (a) homepage,
or (b) book reviews & the latter are pretty dated ... it didn't seem entirely
to merit a heirarchical crumb trail taking up a line or two of nav.

But that meant from the book review pages there was no way to get back to the
homepage.

So:

(a) Add edinburgh.pm.org to the top nav links

(b) Make the banner a hotlink back to the homepage.